Move search functions
Follow up from #5798
